Document type-only constructs
