+++ docs/api/store/azure.md
@@ -1,13 +1,13 @@
 # Microsoft Azure
 
 ::: obstore.store.AzureStore
-::: obstore.store.AzureAccessKey
 ::: obstore.store.AzureConfigInput
     options:
         show_if_no_docstring: true
 ::: obstore.store.AzureConfig
     options:
         show_if_no_docstring: true
+::: obstore.store.AzureAccessKey
 ::: obstore.store.AzureSASToken
 ::: obstore.store.AzureBearerToken
 ::: obstore.store.AzureCredential
