+++ obstore/python/obstore/__sign.py
@@ -1,0 +1,14 @@
+from typing import Literal, TypeAlias
+
+HTTP_METHOD: TypeAlias = Literal[
+    "GET",
+    "PUT",
+    "POST",
+    "HEAD",
+    "PATCH",
+    "TRACE",
+    "DELETE",
+    "OPTIONS",
+    "CONNECT",
+]
+"""Allowed HTTP Methods for signing."""
