+++ obstore/python/obstore/_attributes.py
